refactor: create useCreateAlarmReducer hook
