+++ src/components/alarmAdmin/CreateAlarmModal.tsx
@@ -13,6 +13,7 @@ import ModalFooter from '@/components/common/modal/ModalFooter';
 import ModalHeader from '@/components/common/modal/ModalHeader';
 import OptionTemplate from '@/components/common/OptionTemplate';
 import Selector from '@/components/common/Selector';
+import { useCreateAlarmReducer } from '@/hooks/useCreateAlarmReducer';
 import { currentGenerationState } from '@/recoil/atom';
 import { postNewAlarm } from '@/services/api/alarm';
 import { readPlaygroundId, TARGET_USER_LIST } from '@/utils/alarm';
@@ -23,54 +24,10 @@ interface Props {
   alarmId?: number;
 }
 
-type Action =
-  | { type: 'SET_PART'; payload: string | null }
-  | { type: 'SET_IS_ACTIVE'; payload: boolean | null }
-  | { type: 'SET_TARGET_LIST'; payload: string[] | null }
-  | { type: 'SET_TITLE'; payload: string }
-  | { type: 'SET_CONTENT'; payload: string }
-  | { type: 'SET_LINK'; payload: string | null };
-
-const postAlarmReducer = (draft: PostAlarmData, action: Action): void => {
-  switch (action.type) {
-    case 'SET_PART':
-      draft.part = action.payload;
-      break;
-    case 'SET_IS_ACTIVE':
-      draft.isActive = action.payload;
-      break;
-    case 'SET_TARGET_LIST':
-      draft.targetList = action.payload;
-      break;
-    case 'SET_TITLE':
-      draft.title = action.payload;
-      break;
-    case 'SET_CONTENT':
-      draft.content = action.payload;
-      break;
-    case 'SET_LINK':
-      draft.link = action.payload;
-      break;
-    default:
-      throw new Error('Unknown action type');
-  }
-};
-
-const initialState: PostAlarmData = {
-  target: '활동 회원',
-  part: '발송 파트',
-  isActive: true,
-  targetList: null,
-  title: '',
-  content: '',
-  linkType: '첨부 안함',
-  link: null,
-};
-
 function CreateAlarmModal(props: Props) {
   const { onClose, alarmId } = props;
 
-  const [state, dispatch] = useImmerReducer(postAlarmReducer, initialState);
+  const { state, dispatch } = useCreateAlarmReducer();
 
   const [dropdownVisibility, setDropdownVisibility] = useState({
     part: false,
